requirements.txt
nyuki>=0.11,<0.12
git+https://github.com/aio-libs/aiohttp_cors.git@v0.3.0
docker-py>=1.7,<1.8